[change:mill-py] streamline async and sync method

See merge request qpointz/qpointz!109

+++ mill/core/src/main/java/io/qpointz/mill/types/conversion/SqlDateToEpochDayConverter.java
@@ -3,17 +3,17 @@ package io.qpointz.mill.types.conversion;
 import java.sql.Date;
 import java.time.LocalDate;
 
-public class SqlDateToEpochDayConverter implements ValueConverter<Date, Integer> {
+public class SqlDateToEpochDayConverter implements ValueConverter<Date, Long> {
 
     public static SqlDateToEpochDayConverter DEFAULT = new SqlDateToEpochDayConverter();
 
     @Override
-    public Integer to(Date value) {
-        return (int)value.toLocalDate().toEpochDay();
+    public Long to(Date value) {
+        return value.toLocalDate().toEpochDay();
     }
 
     @Override
-    public Date from(Integer value) {
+    public Date from(Long value) {
        return Date.valueOf(LocalDate.ofEpochDay(value));
     }
